added gitIgnore, separated functions
