+++ .gitignore
@@ -1,0 +1,2 @@
+/base/_pycache_
+/base/views/_pycache_
